+++ src/server/server_ProtocolTranslator.cpp
@@ -20,7 +20,7 @@ IEvent* ProtocolTranslator::translate(Player* player, char code){
   return NULL;
 }
 
-IEvent* ProtocolTranslator::translate(Player* player, char code, int x, int y){
+IEvent* ProtocolTranslator::translate(Player* player, char code, double x, double y){
   switch (code) {
     case PROTOCOL_TELETRANSPORTATION: return new TeletransportationEvent(player, x, y);
     case PROTOCOL_RADIOCONTROLLED: return new RadiocontrolledEvent(player, x, y);
