+++ src/server/server_RadiocontrolledNotification.h
@@ -7,13 +7,13 @@ class RadiocontrolledNotification : public INotification {
 private:
     ClientHandler* client;
     size_t id;
-    int x;
-    int y;
+    double x;
+    double y;
 
     RadiocontrolledNotification(const RadiocontrolledNotification&) = delete;
     RadiocontrolledNotification& operator=(const RadiocontrolledNotification&) = delete;
 public:
-    RadiocontrolledNotification(ClientHandler* client, size_t id, int x, int y);
+    RadiocontrolledNotification(ClientHandler* client, size_t id, double x, double y);
 
     void run();
